refactor: move RecentSearchService into search package
